+++ caliper/managers/pypi.py
@@ -15,8 +15,10 @@ class PypiManager(ManagerBase):
     name = "pypi"
     baseurl = "https://pypi.python.org/pypi"
 
-    def get_package_metadata(self, name=None):
-        """Given a package name, retrieve it's metadata from pypi"""
+    def do_metadata_request(self, name=None):
+        """A separate, shared function to retrieve package metadata without
+        doing any custom filtering.
+        """
         name = name or self.package_name
         if not name:
             raise ValueError("A package name is required.")
@@ -24,14 +26,25 @@ class PypiManager(ManagerBase):
         url = "%s/%s/json" % (self.baseurl, name)
         self.metadata = do_request(url)
 
-        # Note that release[0] can be for any architecture, etc.
-        # The indexing appears consisent within a package, so OK for now
+    @property
+    def releases(self):
+        if not self.metadata:
+            self.do_metadata_request()
+        return self.metadata.get("releases", {})
+
+    def get_package_metadata(self, name=None, arch=None, python_version=None):
+        """Given a package name, retrieve it's metadata from pypi. Given an arch
+        regex and python version, we look for a particular architecture. Otherwise
+        the choices are a bit random.
+        """
+        # Note that without specifying an arch and python version, the
+        # architecture returned can be fairly random.
 
         # Parse metadata into simplified version of spack package schema
-        for version, releases in self.metadata.get("releases", {}).items():
+        for version, releases in self.releases.items():
 
             # Find an appropriate linux/unix flavor release to extract
-            release = self.find_release(releases)
+            release = self.find_release(releases, arch, python_version)
 
             # Some releases can be empty, skip
             if not releases or not release:
@@ -55,10 +68,41 @@ class PypiManager(ManagerBase):
         logger.info("Found %s versions for %s" % (len(self._specs), name))
         return self._specs
 
-    def find_release(self, releases):
+    def get_python_versions(self):
+        """Given a list of releases (or the default) return a list of pep
+        Python versions (e.g., cp38)
+        """
+        python_versions = set()
+        for version, releases in self.releases.items():
+            [
+                python_versions.add(r["python_version"])
+                for r in releases
+                if r["python_version"]
+            ]
+        return python_versions
+
+    def find_release(self, releases=None, arch=None, python_version=None):
         """Given a list of releases, find one that we can extract"""
         filename = None
+        releases = releases or self.releases
+
+        if arch:
+            releases = [r for r in releases if re.search(arch, r["filename"])]
+        if python_version:
+            releases = [
+                r for r in releases if re.search("cp%s" % python_version, r["filename"])
+            ]
+
         for release in releases:
             if re.search("(tar[.]gz|[.]whl)", release["url"]):
                 filename = release
         return filename
+
+    def filter_releases(self, regex, search_field="filename"):
+        """Given a regular expression, filter releases down to smaller list"""
+        filtered = {}
+        for version, releases in self.releases.items():
+            filtered[version] = [
+                r for r in releases if re.search(regex, r.get(search_field, ""))
+            ]
+        return filtered